Split python vs non-python kernel search
